requirements.txt
pandas
numpy
scikit-learn
joblib
seaborn
matplotlib
python-nmap
pyfiglet
